Add RuneLite license back into files which use it
Also split renderer out into runelite-derived part and original part
