+++ src/main/kotlin/controllers/worldRenderer/Renderer.kt
@@ -1,11 +1,8 @@
 package controllers.worldRenderer
 
 import cache.LocationType
-import cache.utils.ColorPalette
 import controllers.worldRenderer.helpers.AlphaMode
 import controllers.worldRenderer.helpers.Animator
-import controllers.worldRenderer.helpers.AntiAliasingMode
-import controllers.worldRenderer.helpers.GpuFloatBuffer
 import controllers.worldRenderer.shaders.Shader
 import controllers.worldRenderer.shaders.ShaderException
 import models.DebugOptionsModel
@@ -14,67 +11,35 @@ import models.config.ConfigOptions
 import models.scene.REGION_SIZE
 import models.scene.Scene
 import models.scene.SceneTile
-import org.joml.Matrix4f
 import org.lwjgl.BufferUtils
 import org.lwjgl.opengl.GL
 import org.lwjgl.opengl.GL.createCapabilities
 import org.lwjgl.opengl.GL11C.GL_BLEND
 import org.lwjgl.opengl.GL11C.GL_COLOR_BUFFER_BIT
-import org.lwjgl.opengl.GL11C.GL_DEPTH_BUFFER_BIT
 import org.lwjgl.opengl.GL11C.GL_DEPTH_TEST
 import org.lwjgl.opengl.GL11C.GL_LEQUAL
 import org.lwjgl.opengl.GL11C.GL_NEAREST
 import org.lwjgl.opengl.GL11C.GL_ONE
 import org.lwjgl.opengl.GL11C.GL_ONE_MINUS_SRC_ALPHA
-import org.lwjgl.opengl.GL11C.GL_RGBA
 import org.lwjgl.opengl.GL11C.GL_SRC_ALPHA
-import org.lwjgl.opengl.GL11C.glClear
-import org.lwjgl.opengl.GL11C.glClearColor
 import org.lwjgl.opengl.GL11C.glDepthFunc
 import org.lwjgl.opengl.GL11C.glDepthRange
 import org.lwjgl.opengl.GL11C.glDisable
 import org.lwjgl.opengl.GL11C.glEnable
-import org.lwjgl.opengl.GL11C.glGetInteger
 import org.lwjgl.opengl.GL11C.glViewport
-import org.lwjgl.opengl.GL13C.GL_MULTISAMPLE
-import org.lwjgl.opengl.GL14C.GL_DEPTH_COMPONENT16
 import org.lwjgl.opengl.GL14C.glBlendFuncSeparate
-import org.lwjgl.opengl.GL15C.GL_DYNAMIC_DRAW
-import org.lwjgl.opengl.GL15C.glBindBuffer
-import org.lwjgl.opengl.GL15C.glBufferData
-import org.lwjgl.opengl.GL15C.glBufferSubData
-import org.lwjgl.opengl.GL15C.glGenBuffers
+import org.lwjgl.opengl.GL20C
 import org.lwjgl.opengl.GL20C.glDeleteProgram
 import org.lwjgl.opengl.GL20C.glDrawBuffers
 import org.lwjgl.opengl.GL20C.glGetUniformLocation
-import org.lwjgl.opengl.GL20C.glUniform1f
 import org.lwjgl.opengl.GL20C.glUniform1i
-import org.lwjgl.opengl.GL20C.glUniform2fv
-import org.lwjgl.opengl.GL20C.glUniform2i
-import org.lwjgl.opengl.GL20C.glUniformMatrix4fv
 import org.lwjgl.opengl.GL20C.glUseProgram
 import org.lwjgl.opengl.GL30C.GL_COLOR_ATTACHMENT0
 import org.lwjgl.opengl.GL30C.GL_COLOR_ATTACHMENT1
-import org.lwjgl.opengl.GL30C.GL_DEPTH_ATTACHMENT
 import org.lwjgl.opengl.GL30C.GL_DRAW_FRAMEBUFFER
-import org.lwjgl.opengl.GL30C.GL_FRAMEBUFFER
-import org.lwjgl.opengl.GL30C.GL_MAX_SAMPLES
 import org.lwjgl.opengl.GL30C.GL_READ_FRAMEBUFFER
-import org.lwjgl.opengl.GL30C.GL_RENDERBUFFER
-import org.lwjgl.opengl.GL30C.glBindBufferBase
 import org.lwjgl.opengl.GL30C.glBindFramebuffer
-import org.lwjgl.opengl.GL30C.glBindRenderbuffer
 import org.lwjgl.opengl.GL30C.glBlitFramebuffer
-import org.lwjgl.opengl.GL30C.glDeleteFramebuffers
-import org.lwjgl.opengl.GL30C.glDeleteRenderbuffers
-import org.lwjgl.opengl.GL30C.glFramebufferRenderbuffer
-import org.lwjgl.opengl.GL30C.glGenFramebuffers
-import org.lwjgl.opengl.GL30C.glGenRenderbuffers
-import org.lwjgl.opengl.GL30C.glRenderbufferStorageMultisample
-import org.lwjgl.opengl.GL30C.glUniform1ui
-import org.lwjgl.opengl.GL31C.GL_UNIFORM_BUFFER
-import org.lwjgl.opengl.GL31C.glGetUniformBlockIndex
-import org.lwjgl.opengl.GL31C.glUniformBlockBinding
 import org.lwjgl.opengl.GL40C.GL_SAMPLE_SHADING
 import org.lwjgl.opengl.GL40C.glMinSampleShading
 import org.lwjgl.opengl.awt.AWTGLCanvas
@@ -86,9 +51,7 @@ import utils.Utils.isMacOS
 import java.awt.event.ActionListener
 import java.awt.event.ComponentAdapter
 import java.awt.event.ComponentEvent
-import java.nio.ByteBuffer
 import java.util.concurrent.ConcurrentLinkedQueue
-import kotlin.math.min
 import kotlin.math.tan
 
 class Renderer(
@@ -99,6 +62,8 @@ class Renderer(
     private val configOptions: ConfigOptions,
     private val frameRateModel: FrameRateModel,
     private val debugOptionsModel: DebugOptionsModel,
+) : RuneliteRenderer(
+    antiAliasingMode = configOptions.antiAliasing.value.get()
 ) {
     enum class PreferredPriorityRenderer(val humanReadableName: String, val factory: () -> PriorityRenderer) {
         GLSL("GLSL Compute-based priority renderer", { GLSLPriorityRenderer() }),
@@ -106,8 +71,6 @@ class Renderer(
     }
 
     private val logger = LoggerFactory.getLogger(Renderer::class.java)
-
-    var antiAliasingMode: AntiAliasingMode = configOptions.antiAliasing.value.get()
     var priorityRendererPref: PreferredPriorityRenderer = configOptions.priorityRenderer.value.get()
         set(value) {
             if (field != value) {
@@ -116,36 +79,16 @@ class Renderer(
             }
         }
 
-    private var glProgram = 0
-
-    private var fboSceneHandle = -1
-    private var rboSceneHandle = -1
-    private var rboSceneDepthBuffer = -1
-
     private var textureArrayId = 0
-    private var uniformBufferId = 0
-    private val uniformBuffer: ByteBuffer = BufferUtils.createByteBuffer(5 * Float.SIZE_BYTES + (3 + 1) * Int.SIZE_BYTES)
-    private val textureOffsets = FloatArray(256)
 
     private lateinit var priorityRenderer: PriorityRenderer
 
     // Uniforms
-    private var uniDrawDistance = 0
-    private var uniViewProjectionMatrix = 0
-    private var uniBrightness = 0
-    private var uniTextures = 0
-    private var uniTextureOffsets = 0
-    private var uniBlockMain = 0
-    private var uniSmoothBanding = 0
-    private var uniMouseCoordsId = 0
+    private var uniMouseCoordsId = -1
     private var uniAlphaMode = -1
-    private var uniHashSeed = -1
 
     var canvasWidth = 100
     var canvasHeight = (canvasWidth / 1.3).toInt()
-    private var lastStretchedCanvasWidth = 0
-    private var lastStretchedCanvasHeight = 0
-    private var lastAntiAliasingMode: AntiAliasingMode? = null
 
     private var animator: Animator? = null
     private var glCanvas: AWTGLCanvas? = null
@@ -158,10 +101,6 @@ class Renderer(
         // center camera in viewport
         camera.centerX = canvasWidth / 2
         camera.centerY = canvasHeight / 2
-
-        fboSceneHandle = -1
-        rboSceneHandle = -1
-        rboSceneDepthBuffer = -1
 
         val glData = GLData()
         glData.majorVersion = 3
@@ -204,9 +143,6 @@ class Renderer(
             frameRateModel.notifyNeedFrames()
         }
 
-        lastStretchedCanvasHeight = -1
-        lastStretchedCanvasWidth = -1
-        lastAntiAliasingMode = null
         textureArrayId = -1
 
         scene.sceneChangeListeners.add(
@@ -261,10 +197,10 @@ class Renderer(
         animator?.start()
     }
 
-    fun stop() {
+    override fun stop() {
         animator!!.stopWith {
+            super.stop()
             shutdownProgram()
-            shutdownAAFbo()
             priorityRenderer.destroy()
         }
     }
@@ -364,10 +300,7 @@ class Renderer(
 
         inputHandler.tick(deltaTime)
 
-        // Setup anti-aliasing
-        val aaEnabled = antiAliasingMode !== AntiAliasingMode.DISABLED
-        if (aaEnabled) {
-            glEnable(GL_MULTISAMPLE)
+        if (isAntiAliasingEnabled()) {
             if (GL.getCapabilities().OpenGL40) {
                 if (configOptions.sampleShading.value.get()) {
                     glEnable(GL_SAMPLE_SHADING)
@@ -379,32 +312,14 @@ class Renderer(
                 logger.warn("Cannot use sample shading on OpenGL < 4.0")
                 configOptions.sampleShading.value.set(false)
             }
-            val stretchedCanvasWidth = canvasWidth
-            val stretchedCanvasHeight = canvasHeight
-
-            // Re-create fbo
-            if (lastStretchedCanvasWidth != stretchedCanvasWidth || lastStretchedCanvasHeight != stretchedCanvasHeight || lastAntiAliasingMode !== antiAliasingMode
-            ) {
-                val maxSamples: Int = glGetInteger(GL_MAX_SAMPLES)
-                val samples = min(antiAliasingMode.ordinal, maxSamples)
-                initAAFbo(stretchedCanvasWidth, stretchedCanvasHeight, samples)
-                lastStretchedCanvasWidth = stretchedCanvasWidth
-                lastStretchedCanvasHeight = stretchedCanvasHeight
-            }
-            glBindFramebuffer(GL_DRAW_FRAMEBUFFER, fboSceneHandle)
-        } else {
-            glDisable(GL_MULTISAMPLE)
-            shutdownAAFbo()
         }
-        lastAntiAliasingMode = antiAliasingMode
+        setupAntiAliasing(canvasWidth, canvasHeight)
+
         // TODO: was this necessary?
         val i = BufferUtils.createIntBuffer(2).put(GL_COLOR_ATTACHMENT0).put(GL_COLOR_ATTACHMENT1).flip()
         glDrawBuffers(i)
 
-        // Clear scene
-        val sky = 9493480
-        glClearColor((sky shr 16 and 0xFF) / 255f, (sky shr 8 and 0xFF) / 255f, (sky and 0xFF) / 255f, 1f)
-        glClear(GL_COLOR_BUFFER_BIT or GL_DEPTH_BUFFER_BIT)
+        clearScene()
 
         val clientCycle = ((System.currentTimeMillis() - clientStart) / 20).toInt() // 50 fps
 
@@ -416,62 +331,8 @@ class Renderer(
             textureArrayId = textureManager.initTextureArray()
         }
 //        val textures: Array<TextureDefinition> = textureProvider.getTextureDefinitions()
-        glUseProgram(glProgram)
-        // Brightness happens to also be stored in the texture provider, so we use that
-        glUniform1f(uniBrightness, ColorPalette.BRIGHTNESS_HIGH.toFloat()) // (float) textureProvider.getBrightness());
-        glUniform1i(uniDrawDistance, Constants.MAX_DISTANCE * Constants.LOCAL_TILE_SIZE)
-        glUniform1f(uniSmoothBanding, 1f)
-        glUniform1ui(uniHashSeed, camera.motionTicks)
-        glUniformMatrix4fv(uniViewProjectionMatrix, false, calculateViewProjectionMatrix())
-
-        // This is just for animating!
-//        for (int id = 0; id < textures.length; ++id) {
-//            TextureDefinition texture = textures[id];
-//            if (texture == null) {
-//                continue;
-//            }
-//
-//            textureProvider.load(id); // trips the texture load flag which lets textures animate
-//
-//            textureOffsets[id * 2] = texture.field1782;
-//            textureOffsets[id * 2 + 1] = texture.field1783;
-//        }
-        glUniform2i(uniMouseCoordsId, inputHandler.mouseX, canvasHeight - inputHandler.mouseY)
-
-        // UBO
-        glBindBuffer(GL_UNIFORM_BUFFER, uniformBufferId)
-        uniformBuffer.clear()
-        val doubles = uniformBuffer.asFloatBuffer()
-        doubles
-            .put(camera.yawRads.toFloat())
-            .put(camera.pitchRads.toFloat())
-            .put(camera.cameraX.toFloat()) // x
-            .put(camera.cameraZ.toFloat()) // z
-            .put(camera.cameraY.toFloat()) // y
-        // Ints
-        val doublesPosition = doubles.position() * Float.SIZE_BYTES
-        uniformBuffer.position(doublesPosition)
-        val ints = uniformBuffer.asIntBuffer()
-        ints
-            .put(camera.centerX)
-            .put(camera.centerY)
-            .put(camera.scale)
-            .put(clientCycle) // currFrame
-        val intsPosition = ints.position() * Int.SIZE_BYTES
-        uniformBuffer.position(doublesPosition + intsPosition)
-        uniformBuffer.flip()
-        glBufferSubData(
-            GL_UNIFORM_BUFFER,
-            0,
-            uniformBuffer
-        )
-        glBindBuffer(GL_UNIFORM_BUFFER, 0)
-
-        // Bind uniforms
-        glBindBufferBase(GL_UNIFORM_BUFFER, 0, uniformBufferId)
-        glUniformBlockBinding(glProgram, uniBlockMain, 0)
-        glUniform1i(uniTextures, 1) // texture sampler array is bound to texture1
-        glUniform2fv(uniTextureOffsets, textureOffsets)
+        prepareDrawProgram(camera, canvasWidth, canvasHeight, clientCycle)
+        GL20C.glUniform2i(uniMouseCoordsId, inputHandler.mouseX, canvasHeight - inputHandler.mouseY)
 
         val useBlend = configOptions.alphaMode.value.get() == AlphaMode.BLEND
         if (useBlend) {
@@ -484,15 +345,7 @@ class Renderer(
         if (useBlend) glDisable(GL_BLEND)
         glUseProgram(0)
 
-        if (aaEnabled) {
-            glBindFramebuffer(GL_READ_FRAMEBUFFER, fboSceneHandle)
-            glBindFramebuffer(GL_DRAW_FRAMEBUFFER, 0)
-            glBlitFramebuffer(
-                0, 0, lastStretchedCanvasWidth, lastStretchedCanvasHeight,
-                0, 0, lastStretchedCanvasWidth, lastStretchedCanvasHeight,
-                GL_COLOR_BUFFER_BIT, GL_NEAREST
-            )
-        }
+        blitAntiAliasing()
 
         glBindFramebuffer(GL_READ_FRAMEBUFFER, 0)
         glBindFramebuffer(GL_DRAW_FRAMEBUFFER, 0)
@@ -672,92 +525,14 @@ class Renderer(
         glUseProgram(0)
     }
 
-    private fun initUniforms() {
-        uniViewProjectionMatrix = glGetUniformLocation(glProgram, "viewProjectionMatrix")
-        uniBrightness = glGetUniformLocation(glProgram, "brightness")
-        uniSmoothBanding = glGetUniformLocation(glProgram, "smoothBanding")
-        uniDrawDistance = glGetUniformLocation(glProgram, "drawDistance")
-        uniTextures = glGetUniformLocation(glProgram, "textures")
-        uniTextureOffsets = glGetUniformLocation(glProgram, "textureOffsets")
-        uniBlockMain = glGetUniformBlockIndex(glProgram, "uniforms")
-        uniMouseCoordsId = glGetUniformLocation(glProgram, "mouseCoords")
+    override fun initUniforms() {
+        super.initUniforms()
+        uniMouseCoordsId = GL20C.glGetUniformLocation(glProgram, "mouseCoords")
         uniAlphaMode = glGetUniformLocation(glProgram, "alphaMode")
-        uniHashSeed = glGetUniformLocation(glProgram, "hashSeed")
-    }
-
-    private fun initUniformBuffer() {
-        uniformBufferId = glGenBuffers()
-        glBindBuffer(GL_UNIFORM_BUFFER, uniformBufferId)
-        uniformBuffer.clear()
-        glBufferData(GL_UNIFORM_BUFFER, uniformBuffer, GL_DYNAMIC_DRAW)
-        glBindBuffer(GL_UNIFORM_BUFFER, 0)
-    }
-
-    private fun initAAFbo(width: Int, height: Int, aaSamples: Int) {
-        // Discard old FBO
-        shutdownAAFbo()
-
-        // Create and bind the FBO
-        fboSceneHandle = glGenFramebuffers()
-        glBindFramebuffer(GL_FRAMEBUFFER, fboSceneHandle)
-
-        // Create color render buffer
-        rboSceneHandle = glGenRenderbuffers()
-        glBindRenderbuffer(GL_RENDERBUFFER, rboSceneHandle)
-        glRenderbufferStorageMultisample(GL_RENDERBUFFER, aaSamples, GL_RGBA, width, height)
-        glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_RENDERBUFFER, rboSceneHandle)
-
-        // Create depth buffer
-        rboSceneDepthBuffer = glGenRenderbuffers()
-        glBindRenderbuffer(GL_RENDERBUFFER, rboSceneDepthBuffer)
-        glRenderbufferStorageMultisample(GL_RENDERBUFFER, aaSamples, GL_DEPTH_COMPONENT16, width, height)
-        glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, rboSceneDepthBuffer)
-
-        // Reset
-        glBindFramebuffer(GL_FRAMEBUFFER, 0)
-        glBindRenderbuffer(GL_RENDERBUFFER, 0)
-    }
-
-    private fun makeProjectionMatrix(width: Float, height: Float, near: Float): FloatArray =
-        floatArrayOf(
-            2 / width, 0f, 0f, 0f,
-            0f, 2 / height, 0f, 0f,
-            0f, 0f, -1f, -1f,
-            0f, 0f, -2 * near, 0f,
-        )
-
-    private fun calculateViewProjectionMatrix(): FloatArray {
-        val projectionMatrix = makeProjectionMatrix(
-            canvasWidth.toFloat(),
-            canvasHeight.toFloat(),
-            50.0f
-        )
-        return Matrix4f()
-            .scale(camera.scale.toFloat(), camera.scale.toFloat(), 1.0f)
-            .mul(Matrix4f(GpuFloatBuffer.allocateDirect(projectionMatrix.size).put(projectionMatrix).flip()))
-            .rotateX((-Math.PI + camera.pitchRads).toFloat())
-            .rotateY(camera.yawRads.toFloat())
-            .translate(-camera.cameraX.toFloat(), -camera.cameraZ.toFloat(), -camera.cameraY.toFloat())
-            .get(projectionMatrix)
     }
 
     private fun shutdownProgram() {
         glDeleteProgram(glProgram)
         glProgram = -1
-    }
-
-    private fun shutdownAAFbo() {
-        if (fboSceneHandle != -1) {
-            glDeleteFramebuffers(fboSceneHandle)
-            fboSceneHandle = -1
-        }
-        if (rboSceneHandle != -1) {
-            glDeleteRenderbuffers(rboSceneHandle)
-            rboSceneHandle = -1
-        }
-        if (rboSceneDepthBuffer != -1) {
-            glDeleteRenderbuffers(rboSceneDepthBuffer)
-            rboSceneDepthBuffer = -1
-        }
     }
 }
